improved us delay

+++ uCNC/src/hal/mcus/mcu.h
@@ -271,7 +271,7 @@ extern "C"
 	 * the maximum allowed delay is 255 us
 	 * */
 #ifndef mcu_delay_us
-	void mcu_delay_us(uint8_t delay);
+	void mcu_delay_us(uint16_t delay);
 #endif
 
 	/**
